+++ Classes/UVArticle.m
@@ -7,11 +7,41 @@
 //
 
 #import "UVArticle.h"
+#import "UVSuggestion.h"
+#import "UVResponseDelegate.h"
+#import "UVSession.h"
+#import "UVClientConfig.h"
+#import "UVForum.h"
 
 @implementation UVArticle
 
 @synthesize question;
 @synthesize answerHTML;
+
++ (void)initialize {
+	[self setDelegate:[[UVResponseDelegate alloc] initWithModelClass:[self class]]];
+	[self setBaseURL:[self siteURL]];
+}
+
++ (NSArray *)getInstantAnswers:(NSString *)query delegate:(id)delegate {
+    NSDictionary *params = [NSDictionary dictionaryWithObjectsAndKeys:
+                            @"3", @"per_page",
+                            [NSString stringWithFormat:@"%d", [UVSession currentSession].clientConfig.forum.forumId], @"forum_id",
+							query, @"query",
+							nil];
+
+    return [self getPath:[self apiPath:@"/instant_answers/search.json"]
+			  withParams:params
+				  target:delegate
+				selector:@selector(didRetrieveInstantAnswers:)];
+}
+
++ (UVBaseModel *)modelForDictionary:(NSDictionary *)dict {
+    NSString *type = [dict objectForKey:@"type"];
+    if ([@"suggestion" isEqualToString:type])
+        return [UVSuggestion modelForDictionary:dict];
+    return [super modelForDictionary:dict];
+}
 
 - (id)initWithDictionary:(NSDictionary *)dict {
     if ((self = [super init])) {