+++ Classes/UVArticleViewController.m
@@ -22,11 +22,16 @@
 
 - (void)loadView {
     [super loadView];
+    [self hideExitButton];
     self.webView = [[[UIWebView alloc] initWithFrame:[self contentFrame]] autorelease];
-    NSString *html = [NSString stringWithFormat:@"<html><head><link rel=\"stylesheet\" type=\"text/css\" href=\"http://cdn.uservoice.com/stylesheets/vendor/typeset.css\"/></head><body class=\"typeset\" style=\"font-family: sans-serif\"><h3>%@</h3>%@</body></html>", article.question, article.answerHTML];
+    NSString *html = [NSString stringWithFormat:@"<html><head><link rel=\"stylesheet\" type=\"text/css\" href=\"http://cdn.uservoice.com/stylesheets/vendor/typeset.css\"/></head><body class=\"typeset\" style=\"font-family: sans-serif; margin: 1em\"><h3>%@</h3>%@</body></html>", article.question, article.answerHTML];
     NSLog(@"%@", html);
     [self.webView loadHTMLString:html baseURL:nil];
     self.view = self.webView;
+}
+
+- (NSString *)backButtonTitle {
+    return @"Back";
 }
 
 - (void)dealloc {