+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 ## master ##
 
-* Remove sign out button if the user is signed in programmatically
-* Fix a cosmetic issue with custom fields on the iPad
 * Add top articles and suggestions to welcome screen
 * Make ordering of suggestions match feedback site default (Hot Ideas, New Ideas, or Top Ideas)
+* Add instant answers to contact form
+
+## 1.1.1 ##
+
+* Remove sign out button if the user is signed in programmatically
+* Fix a cosmetic issue with custom fields on the iPad
 
 ## 1.1.0 ##